pyproject.toml
[project]
name = "bug-5702927"
version = "0.1.0"
description = "Testing https://github.com/Azure/azure-sdk-for-python/issues/42337"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-monitor-opentelemetry>=1.6.13",
    "pip-system-certs>=5.2",
]
